store: split out assertion support to store_asserts.go
